move the tests to it's own file
